Fix broken link of section 2 index.
Original file name link type using align number which 1 to 01 if max serial
number was 10 and so on. Although this type was tidy and pretty
but with serious complicated transfer for further contents adding
in the furture, because that org mode raw manue was typesetting as
the current modified thing, and do along with it was good I think.

On branch master
Changes to be committed:
	modified:   org/2_Document-structure/2_Document-structure.org